Refactor: split code into three new components
