+++ src/utils/trimTitle.ts
@@ -1,0 +1,3 @@
+export function trimTitle(text: string) {
+  return text.replace(/\s+/g, ' ').trim();
+}
